requirements.txt
Flask==1.1.2
python-dotenv==0.15.0
Flask-WTF==0.14.3
Flask-SQLAlchemy==2.4.4
WTForms~=2.3.3
requests~=2.24.0
Flask-Bootstrap==3.3.7.1
python-dateutil~=2.8.1
